+++ src/Api/V5/Contract/CpcVideoAdBuilderAdAdd.php
@@ -8,7 +8,11 @@ namespace Biplane\YandexDirect\Api\V5\Contract;
 class CpcVideoAdBuilderAdAdd extends AdBuilderAdAddBase
 {
 
-    protected $Href = null;
+//    Can be omit.
+//    protected $Href = null;
+
+//    Can be omit.
+//    protected $TurboPageId = null;
 
     /**
      * Creates a new instance of CpcVideoAdBuilderAdAdd.
@@ -23,22 +27,45 @@ class CpcVideoAdBuilderAdAdd extends AdBuilderAdAddBase
     /**
      * Gets Href.
      *
-     * @return string
+     * @return string|null
      */
     public function getHref()
     {
-        return $this->Href;
+        return isset($this->Href) ? $this->Href : null;
     }
 
     /**
      * Sets Href.
      *
-     * @param string $value
+     * @param string|null $value
      * @return $this
      */
-    public function setHref($value)
+    public function setHref($value = null)
     {
         $this->Href = $value;
+
+        return $this;
+    }
+
+    /**
+     * Gets TurboPageId.
+     *
+     * @return int|null
+     */
+    public function getTurboPageId()
+    {
+        return isset($this->TurboPageId) ? $this->TurboPageId : null;
+    }
+
+    /**
+     * Sets TurboPageId.
+     *
+     * @param int|null $value
+     * @return $this
+     */
+    public function setTurboPageId($value = null)
+    {
+        $this->TurboPageId = $value;
 
         return $this;
     }