+++ src/Api/V5/Contract/CpcVideoAdBuilderAdGet.php
@@ -11,6 +11,12 @@ class CpcVideoAdBuilderAdGet extends AdBuilderAdGetBase
 //    Can be omit.
 //    protected $Href = null;
 
+//    Can be omit.
+//    protected $TurboPageId = null;
+
+//    Can be omit.
+//    protected $TurboPageModeration = null;
+
     /**
      * Creates a new instance of CpcVideoAdBuilderAdGet.
      *
@@ -40,6 +46,52 @@ class CpcVideoAdBuilderAdGet extends AdBuilderAdGetBase
     public function setHref($value = null)
     {
         $this->Href = $value;
+
+        return $this;
+    }
+
+    /**
+     * Gets TurboPageId.
+     *
+     * @return int|null
+     */
+    public function getTurboPageId()
+    {
+        return isset($this->TurboPageId) ? $this->TurboPageId : null;
+    }
+
+    /**
+     * Sets TurboPageId.
+     *
+     * @param int|null $value
+     * @return $this
+     */
+    public function setTurboPageId($value = null)
+    {
+        $this->TurboPageId = $value;
+
+        return $this;
+    }
+
+    /**
+     * Gets TurboPageModeration.
+     *
+     * @return ExtensionModeration|null
+     */
+    public function getTurboPageModeration()
+    {
+        return isset($this->TurboPageModeration) ? $this->TurboPageModeration : null;
+    }
+
+    /**
+     * Sets TurboPageModeration.
+     *
+     * @param ExtensionModeration|null $value
+     * @return $this
+     */
+    public function setTurboPageModeration(ExtensionModeration $value = null)
+    {
+        $this->TurboPageModeration = $value;
 
         return $this;
     }