+++ src/Api/V5/Contract/CpcVideoAdBuilderAdUpdate.php
@@ -11,6 +11,9 @@ class CpcVideoAdBuilderAdUpdate extends AdBuilderAdUpdateBase
 //    Can be omit.
 //    protected $Href = null;
 
+//    Can be omit.
+//    protected $TurboPageId = null;
+
     /**
      * Creates a new instance of CpcVideoAdBuilderAdUpdate.
      *
@@ -40,6 +43,29 @@ class CpcVideoAdBuilderAdUpdate extends AdBuilderAdUpdateBase
     public function setHref($value = null)
     {
         $this->Href = $value;
+
+        return $this;
+    }
+
+    /**
+     * Gets TurboPageId.
+     *
+     * @return int|null
+     */
+    public function getTurboPageId()
+    {
+        return isset($this->TurboPageId) ? $this->TurboPageId : null;
+    }
+
+    /**
+     * Sets TurboPageId.
+     *
+     * @param int|null $value
+     * @return $this
+     */
+    public function setTurboPageId($value = null)
+    {
+        $this->TurboPageId = $value;
 
         return $this;
     }