+++ src/Api/V5/Contract/CpmBannerAdBuilderAdGet.php
@@ -14,6 +14,12 @@ class CpmBannerAdBuilderAdGet extends AdBuilderAdGetBase
 //    Can be omit.
 //    protected $TrackingPixels = null;
 
+//    Can be omit.
+//    protected $TurboPageId = null;
+
+//    Can be omit.
+//    protected $TurboPageModeration = null;
+
     /**
      * Creates a new instance of CpmBannerAdBuilderAdGet.
      *
@@ -66,6 +72,52 @@ class CpmBannerAdBuilderAdGet extends AdBuilderAdGetBase
     public function setTrackingPixels(array $value = null)
     {
         $this->TrackingPixels = $value;
+
+        return $this;
+    }
+
+    /**
+     * Gets TurboPageId.
+     *
+     * @return int|null
+     */
+    public function getTurboPageId()
+    {
+        return isset($this->TurboPageId) ? $this->TurboPageId : null;
+    }
+
+    /**
+     * Sets TurboPageId.
+     *
+     * @param int|null $value
+     * @return $this
+     */
+    public function setTurboPageId($value = null)
+    {
+        $this->TurboPageId = $value;
+
+        return $this;
+    }
+
+    /**
+     * Gets TurboPageModeration.
+     *
+     * @return ExtensionModeration|null
+     */
+    public function getTurboPageModeration()
+    {
+        return isset($this->TurboPageModeration) ? $this->TurboPageModeration : null;
+    }
+
+    /**
+     * Sets TurboPageModeration.
+     *
+     * @param ExtensionModeration|null $value
+     * @return $this
+     */
+    public function setTurboPageModeration(ExtensionModeration $value = null)
+    {
+        $this->TurboPageModeration = $value;
 
         return $this;
     }