+++ src/Api/V5/Contract/SitelinkAddItem.php
@@ -5,23 +5,51 @@ namespace Biplane\YandexDirect\Api\V5\Contract;
 /**
  * Auto-generated code.
  */
-class TextImageAdUpdate extends ImageAdUpdateBase
+class SitelinkAddItem
 {
+
+    protected $Title = null;
 
 //    Can be omit.
 //    protected $Href = null;
 
 //    Can be omit.
+//    protected $Description = null;
+
+//    Can be omit.
 //    protected $TurboPageId = null;
 
     /**
-     * Creates a new instance of TextImageAdUpdate.
+     * Creates a new instance of SitelinkAddItem.
      *
      * @return self
      */
     public static function create()
     {
         return new self();
+    }
+
+    /**
+     * Gets Title.
+     *
+     * @return string
+     */
+    public function getTitle()
+    {
+        return $this->Title;
+    }
+
+    /**
+     * Sets Title.
+     *
+     * @param string $value
+     * @return $this
+     */
+    public function setTitle($value)
+    {
+        $this->Title = $value;
+
+        return $this;
     }
 
     /**
@@ -43,6 +71,29 @@ class TextImageAdUpdate extends ImageAdUpdateBase
     public function setHref($value = null)
     {
         $this->Href = $value;
+
+        return $this;
+    }
+
+    /**
+     * Gets Description.
+     *
+     * @return string|null
+     */
+    public function getDescription()
+    {
+        return isset($this->Description) ? $this->Description : null;
+    }
+
+    /**
+     * Sets Description.
+     *
+     * @param string|null $value
+     * @return $this
+     */
+    public function setDescription($value = null)
+    {
+        $this->Description = $value;
 
         return $this;
     }