+++ src/Api/V5/Contract/SitelinkFieldEnum.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Biplane\YandexDirect\Api\V5\Contract;
+
+/**
+ * Auto-generated code.
+ */
+class SitelinkFieldEnum
+{
+
+    const TITLE = 'Title';
+
+    const HREF = 'Href';
+
+    const DESCRIPTION = 'Description';
+
+    const TURBO_PAGE_ID = 'TurboPageId';
+
+
+}
+