+++ src/Api/V5/Contract/TextAdBuilderAdFieldEnum.php
@@ -12,6 +12,10 @@ class TextAdBuilderAdFieldEnum
 
     const HREF = 'Href';
 
+    const TURBO_PAGE_ID = 'TurboPageId';
+
+    const TURBO_PAGE_MODERATION = 'TurboPageModeration';
+
 
 }